scripts/requirements.txt
httpx==0.27.0
# Python 3.13 and below
psycopg2-binary==2.9.9; python_version < "3.14"
# Python 3.14+
psycopg[binary]>=3.2.0,<4; python_version >= "3.14"
pgvector==0.3.2
google-generativeai==0.7.2
# Python 3.12 and below
numpy==1.26.4; python_version < "3.13"
# Python 3.13+
numpy>=2.1.0,<3; python_version >= "3.13"
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
mediapipe>=0.10.30,<0.11
